Support creating InstanceMetadata from env
In Ev2 deployments the deployer is running in Azure Container Instance
(ACI) and does _not_ have access to the instance metadata we have used
to date.  But we have the info to populate env vars.  So we will do that
and attempt to use env vars first.  If we get everything we need, log
what happened and do not use metadata endpoint.  Else fallback to
existing behavior.